Filter with rolling median

Using a rolling median, filter the NDVI time series for each id.

filter_roll(DT, window = 3L, id = "id", method = "median")

Arguments

DT
data.table of NDVI time series

window
window size. default is 3.

id
id column. default is 'id'. See details.

method
median. no other options yet. let me know if you are looking for something else.

Value

filtered data.table with appended 'rolled' column of each id's rolling median, filtered NDVI time series.

Details

The id argument is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis.

See also

Examples

# Load data.table
library(data.table)

# Read example data
ndvi <- fread(system.file("extdata", "ndvi.csv", package = "irg"))

filter_qa(ndvi, qa = 'SummaryQA', good = c(0, 1))
filter_winter(ndvi, probs = 0.025, limits = c(60L, 300L), doy = 'DayOfYear', id = 'id')
#>       id   yr DayOfYear  NDVI SummaryQA filtered winter
#>    1:  1 2002         3 -1367         3     4099   4099
#>    2:  2 2002        14  -304         3     5382   5382
#>    3:  3 2002         1   374         2     3702   3702
#>    4:  4 2002        15   635         3     5180   5180
#>    5:  5 2002         9   685         2     4621   4621
#>   ---                                                  
#> 1261:  1 2012       353   151         2     4099   4099
#> 1262:  2 2012       356   330         2     5382   5382
#> 1263:  3 2012       356   560         2     3702   3702
#> 1264:  4 2012       356  1720         2     5180   5180
#> 1265:  5 2012       356  2689         2     4621   4621
filter_roll(ndvi, window = 3L, id = 'id')
#>       id   yr DayOfYear  NDVI SummaryQA filtered winter rolled
#>    1:  1 2002         3 -1367         3     4099   4099   4099
#>    2:  2 2002        14  -304         3     5382   5382   5382
#>    3:  3 2002         1   374         2     3702   3702   3702
#>    4:  4 2002        15   635         3     5180   5180   5180
#>    5:  5 2002         9   685         2     4621   4621   4621
#>   ---                                                         
#> 1261:  1 2012       353   151         2     4099   4099   4099
#> 1262:  2 2012       356   330         2     5382   5382   5382
#> 1263:  3 2012       356   560         2     3702   3702   3702
#> 1264:  4 2012       356  1720         2     5180   5180   5180
#> 1265:  5 2012       356  2689         2     4621   4621   4621