IRG

Wrapper function for one step IRG calculation. Only defaults.

irg(DT)

Arguments

DT

data.table of NDVI time series

Value

Extended data.table 'irg' column of instantaneous rate of green-up calculated for each day of the year, for each individual and year.

Details

data.table must have columns:

'id' - individual identifier

'yr' - year of observation

'NDVI' - NDVI value

'DayOfYear' - day of year/julian day of observation

'SummaryQA' - summary quality value for each sample (provided by MODIS)

See also

Examples

# Load data.table
library(data.table)

# Read in example data
ndvi <- fread(system.file("extdata", "sampled-ndvi-MODIS-MOD13Q1.csv", package = "irg"))

# Calculate IRG for each day of the year and individual
out <- irg(ndvi)