Model starting parameters

Try guessing starting parameters for model_params and model_ndvi.

model_start(DT, id = "id", year = "yr")

Arguments

DT

filtered and scaled data.table of NDVI time series. Expects columns 'scaled' and 't' are present.

id

id column. default is 'id'. See details.

year

year column name. default is 'yr'.

Value

The input DT data.table appended with xmidS_start and xmidA_start columns. Note - we curently do not attempt to guess appropriate starting values for scalS and scalA.

Details

The id argument is used to split between sampling units. This may be a point id, polygon id, pixel id, etc. depending on your analysis. This should match the id provided to filtering functions.

See also

Examples

# Load data.table
library(data.table)

# Read in example data
ndvi <- fread(system.file("extdata", "sampled-ndvi-MODIS-MOD13Q1.csv", package = "irg"))

# Filter and scale NDVI time series
filter_ndvi(ndvi)
#>      id   NDVI SummaryQA DayOfYear   yr  filtered    winter    rolled       top
#>   1:  0 0.1864         3        11 2015 0.3076500 0.3076500 0.3076500 0.8735000
#>   2:  1 0.0541         2         3 2015 0.3163400 0.3163400 0.3163400 0.8632175
#>   3:  2 0.1781         3        11 2015 0.2649875 0.2649875 0.2649875 0.8707500
#>   4:  3 0.1024         2         5 2015 0.2301750 0.2301750 0.2301750 0.8635525
#>   5:  4 0.0898         2         3 2015 0.2177150 0.2177150 0.2177150 0.8476000
#>  ---                                                                           
#> 801:  2 0.1179         2       364 2019 0.2649875 0.2649875 0.2649875 0.8707500
#> 802:  3 0.0789         2       364 2019 0.2301750 0.2301750 0.2301750 0.8635525
#> 803:  4 0.1572         2       364 2019 0.2177150 0.2177150 0.2177150 0.8476000
#> 804:  5 0.0763         2       364 2019 0.3163400 0.3163400 0.3163400 0.8632175
#> 805:  6 0.1197         2       362 2019 0.3149325 0.3149325 0.3149325 0.8632000
scale_doy(ndvi)
#>      id   NDVI SummaryQA DayOfYear   yr  filtered    winter    rolled       top
#>   1:  0 0.1864         3        11 2015 0.3076500 0.3076500 0.3076500 0.8735000
#>   2:  1 0.0541         2         3 2015 0.3163400 0.3163400 0.3163400 0.8632175
#>   3:  2 0.1781         3        11 2015 0.2649875 0.2649875 0.2649875 0.8707500
#>   4:  3 0.1024         2         5 2015 0.2301750 0.2301750 0.2301750 0.8635525
#>   5:  4 0.0898         2         3 2015 0.2177150 0.2177150 0.2177150 0.8476000
#>  ---                                                                           
#> 801:  2 0.1179         2       364 2019 0.2649875 0.2649875 0.2649875 0.8707500
#> 802:  3 0.0789         2       364 2019 0.2301750 0.2301750 0.2301750 0.8635525
#> 803:  4 0.1572         2       364 2019 0.2177150 0.2177150 0.2177150 0.8476000
#> 804:  5 0.0763         2       364 2019 0.3163400 0.3163400 0.3163400 0.8632175
#> 805:  6 0.1197         2       362 2019 0.3149325 0.3149325 0.3149325 0.8632000
#>                t
#>   1: 0.027397260
#>   2: 0.005479452
#>   3: 0.027397260
#>   4: 0.010958904
#>   5: 0.005479452
#>  ---            
#> 801: 0.994520548
#> 802: 0.994520548
#> 803: 0.994520548
#> 804: 0.994520548
#> 805: 0.989041096
scale_ndvi(ndvi)
#>      id   NDVI SummaryQA DayOfYear   yr  filtered    winter    rolled       top
#>   1:  0 0.1864         3        11 2015 0.3076500 0.3076500 0.3076500 0.8735000
#>   2:  1 0.0541         2         3 2015 0.3163400 0.3163400 0.3163400 0.8632175
#>   3:  2 0.1781         3        11 2015 0.2649875 0.2649875 0.2649875 0.8707500
#>   4:  3 0.1024         2         5 2015 0.2301750 0.2301750 0.2301750 0.8635525
#>   5:  4 0.0898         2         3 2015 0.2177150 0.2177150 0.2177150 0.8476000
#>  ---                                                                           
#> 801:  2 0.1179         2       364 2019 0.2649875 0.2649875 0.2649875 0.8707500
#> 802:  3 0.0789         2       364 2019 0.2301750 0.2301750 0.2301750 0.8635525
#> 803:  4 0.1572         2       364 2019 0.2177150 0.2177150 0.2177150 0.8476000
#> 804:  5 0.0763         2       364 2019 0.3163400 0.3163400 0.3163400 0.8632175
#> 805:  6 0.1197         2       362 2019 0.3149325 0.3149325 0.3149325 0.8632000
#>                t scaled
#>   1: 0.027397260      0
#>   2: 0.005479452      0
#>   3: 0.027397260      0
#>   4: 0.010958904      0
#>   5: 0.005479452      0
#>  ---                   
#> 801: 0.994520548      0
#> 802: 0.994520548      0
#> 803: 0.994520548      0
#> 804: 0.994520548      0
#> 805: 0.989041096      0

# Guess starting parameters for xmidS and xmidA
model_start(ndvi)
#>      id   NDVI SummaryQA DayOfYear   yr filtered    winter rolled       top
#>   1:  0 0.0791         2        61 2016       NA 0.3076500     NA 0.8735000
#>   2:  1 0.0526         2        61 2016       NA 0.3163400     NA 0.8632175
#>   3:  2 0.0707         2        61 2016       NA 0.2649875     NA 0.8707500
#>   4:  3 0.0456         2        61 2016       NA 0.2301750     NA 0.8635525
#>   5:  5 0.0526         2        61 2016       NA 0.3163400     NA 0.8632175
#>  ---                                                                       
#> 801:  1 0.8755         1       225 2015   0.8755 0.3163400 0.8671 0.8632175
#> 802:  2 0.8740         1       225 2015   0.8740 0.2649875 0.8732 0.8707500
#> 803:  3 0.8643         1       225 2015   0.8643 0.2301750 0.8643 0.8635525
#> 804:  5 0.8755         1       225 2015   0.8755 0.3163400 0.8671 0.8632175
#> 805:  6 0.8729         1       225 2015   0.8729 0.3149325 0.8688 0.8632000
#>              t scaled xmidS_start xmidA_start
#>   1: 0.1643836     NA   0.3452055   0.7095890
#>   2: 0.1643836     NA   0.3452055   0.7095890
#>   3: 0.1643836     NA   0.3452055   0.7095890
#>   4: 0.1643836     NA   0.3452055   0.7095890
#>   5: 0.1643836     NA   0.3452055   0.7095890
#>  ---                                         
#> 801: 0.6136986      1   0.4000000   0.7123288
#> 802: 0.6136986      1   0.3808219   0.7123288
#> 803: 0.6136986      1   0.3808219   0.7808219
#> 804: 0.6136986      1   0.4000000   0.7123288
#> 805: 0.6136986      1   0.3808219   0.7808219