Raw Landsat NDVI data

A CSV containing NDVI samples for seven points over ten years (2005-2010). Data extracted using Earth Engine with the example script provided by the use_example_ee_script() function with sensor set to 'Landsat'.

Format

A data.table with 1652 rows and 5 variables:

id - individual identifier

ndvi - sampled NDVI value

mask - mask value, see details below

doy - julian day/day of year of sample

year - year of sample

mask details:

0 - Good data

1 - if QA_PIXEL indicates unwanted pixels OR if QA_RADSAT indicates saturated pixels

2 - if QA_PIXEL indicates unwanted pixels AND if QA_RADSAT indicates saturated pixels

Details

Note: these are the same locations as in the example 'MODIS' data.

Examples

# Load data.table
library(data.table)

# Read example data
ndvi <- fread(system.file('extdata', 'sampled-ndvi-Landsat-LC08-T1-L2.csv', package = 'irg'))