Use an example sampling script for Earth Engine

Provides an example script for use in Earth Engine, as a preceeding step to using the irg package. Use the script to sample NDVI in Earth Engine, then use the irg package to calculate the instantaneous rate of green-up.

use_example_ee_script(sensor = "MODIS", filename = NULL, overwrite = FALSE)

Arguments

sensor

either 'MODIS' or 'Landsat'

filename

file name indicating where to save the example script. Default is './Sample-NDVI-For-IRG-MODIS.js' if sensor is 'MODIS' or 'Sample-NDVI-For-IRG-Landsat.js' if sensor is 'Landsat'.

overwrite

boolean indicating if the file should overwrite existing files. default is FALSE.

Value

use_example_ee_script saves an example NDVI extraction script in the current working directory using a default filename, or if filename is provided, at the location specified.